# Re-enable static cached_op optimization