@marcoabreu 2019-05-11
Was there a test added in the process of resolving the linked issue?