@pengzhao-intel 2019-05-12
> Was there a test added in the process of resolving the linked issue?

This is a performance issue in the gluonNLP so any suggestions for adding performance test on CI?