@pengzhao-intel 2019-05-14
Will merge this PR at the end of today if there is no further comment.